requirements-prod.txt
# ===================================================================
# Production Dependencies
# ===================================================================

# Include core requirements
-r requirements.txt

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Performance & Caching
Flask-Caching==2.1.0
redis==5.0.1

# Security
Flask-Talisman==1.1.0
cryptography==41.0.7

# Monitoring & Error Tracking
sentry-sdk[flask]==1.38.0
prometheus-client==0.19.0

# Rate Limiting
Flask-Limiter==3.5.0

# Health Checks
Flask-HealthCheck==1.0.0

# Compression
Flask-Compress==1.14

# Enhanced Image Processing
scikit-image==0.22.0
scipy==1.11.4

# Background Tasks (if needed)
celery==5.3.4

# Database Connection Pooling
SQLAlchemy-Utils==0.41.1

# Logging
structlog==23.2.0

# Environment Variables
python-decouple==3.8